correcciones de rubocop

+++ config/application.rb
@@ -13,9 +13,11 @@ module Autonoma
     config.load_defaults 5.1
     config.before_configuration do
       env_file = File.join(Rails.root, 'config', 'local_env.yml')
-      YAML.load(File.open(env_file)).each do |key, value|
-        ENV[key.to_s] = value.to_s
-      end if File.exists?(env_file)
+      if File.exist?(env_file)
+        YAML.load(File.open(env_file)).each do |key, value|
+          ENV[key.to_s] = value.to_s
+        end
+      end
     end
 
     config.i18n.load_path += Dir[Rails.root.join('config', 'locales', '**', '*.{rb,yml}').to_s]
@@ -26,15 +28,15 @@ module Autonoma
     config.i18n.default_locale = :"es-CO"
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration should go into files in config/initializers
-    config.assets.precompile += [ 'appviews.css', 'cssanimations.css', 'dashboards.css', 'forms.css', 'gallery.css', 'graphs.css', 'mailbox.css', 'miscellaneous.css', 'pages.css', 'tables.css', 'uielements.css', 'widgets.css', 'commerce.css' ]
-    config.assets.precompile += [ 'appviews.js', 'cssanimations.js', 'dashboards.js', 'forms.js', 'gallery.js', 'graphs.js', 'mailbox.js', 'miscellaneous.js', 'pages.js', 'tables.js', 'uielements.js', 'widgets.js', 'commerce.js', 'metrics.js', 'landing.js' ]
+    config.assets.precompile += ['appviews.css', 'cssanimations.css', 'dashboards.css', 'forms.css', 'gallery.css', 'graphs.css', 'mailbox.css', 'miscellaneous.css', 'pages.css', 'tables.css', 'uielements.css', 'widgets.css', 'commerce.css']
+    config.assets.precompile += ['appviews.js', 'cssanimations.js', 'dashboards.js', 'forms.js', 'gallery.js', 'graphs.js', 'mailbox.js', 'miscellaneous.js', 'pages.js', 'tables.js', 'uielements.js', 'widgets.js', 'commerce.js', 'metrics.js', 'landing.js']
     # -- all .rb files in that directory are automatically loaded.
     config.to_prepare do
-      Devise::SessionsController.layout "session"
-      Devise::PasswordsController.layout "session"
-      Devise::PasswordsController.layout "session"
-      Devise::ConfirmationsController.layout "session"
-      Devise::RegistrationsController.layout proc{ |controller| user_signed_in? ? "application" : "session" }
+      Devise::SessionsController.layout 'session'
+      Devise::PasswordsController.layout 'session'
+      Devise::PasswordsController.layout 'session'
+      Devise::ConfirmationsController.layout 'session'
+      Devise::RegistrationsController.layout proc { |_controller| user_signed_in? ? 'application' : 'session' }
       # Or to configure mailer layout
       # Devise::Mailer.layout "email" # email.haml or email.erb
     end
